requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-api-key==3.1.0
packaging==25.0
sqlparse==0.5.3
